as: Don't 'unlink (NULL)'
Otherwise, when compiled with '-fsanitize=undefined':

    $ echo § | ./nvptx-none-as
    [...]
    nvptx-as.cc:95:12: runtime error: null pointer passed as argument 1, which is declared to never be null

+++ nvptx-as.cc
@@ -91,7 +91,9 @@ fatal_error (const char * cmsgid, ...)
   fprintf (stderr, "\n");
   va_end (ap);
 
-  unlink (outname);
+  if (outname)
+    unlink (outname);
+
   exit (1);
 }
